git-svn-id: http://ownageclan.com/svn/war3source@335 e104397d-def3-435e-8314-01917568762b

+++ extensionsource/war3ext/mytimer.h
@@ -41,7 +41,7 @@ private:
 				for(int i=0;i<size;i++){
 					TimerEntry *te=myvector.at(i);
 					if(te->nexttick<min){
-						min=te->nexttick;
+						min=(float)te->nexttick;
 						minindex=i;
 						readycount++;
 					}
